fix: 새로운 구조의 jdbcTemplate로 리팩터링
