+++ app/src/main/java/com/techcourse/dao/UserHistoryDao.java
@@ -1,6 +1,6 @@
 package com.techcourse.dao;
 
-import com.interface21.jdbc.core.JdbcTemplate2;
+import com.interface21.jdbc.core.JdbcTemplate;
 import com.techcourse.domain.UserHistory;
 import javax.sql.DataSource;
 import org.slf4j.Logger;
@@ -10,13 +10,13 @@ public class UserHistoryDao {
 
     private static final Logger log = LoggerFactory.getLogger(UserHistoryDao.class);
 
-    private final JdbcTemplate2 jdbcTemplate2;
+    private final JdbcTemplate jdbcTemplate2;
 
     public UserHistoryDao(final DataSource dataSource) {
-        jdbcTemplate2 = new JdbcTemplate2(dataSource);
+        jdbcTemplate2 = new JdbcTemplate(dataSource);
     }
 
-    public UserHistoryDao(final JdbcTemplate2 jdbcTemplate) {
+    public UserHistoryDao(final JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate2 = jdbcTemplate;
     }
